Open either edit account info or password depending section depending on button
