+++ src/pages/settings/adminView.jsx
@@ -1,99 +1,16 @@
 import React from 'react';
 import { PropTypes } from 'prop-types';
-import { Container, Row, Col } from 'react-bootstrap';
-import { Link } from 'react-router-dom';
 import styles from './settings.module.css';
+import AccountInformationView from './accountInformationView';
+import PasswordView from './passwordView';
 
 const AdminView = ({ userInfo }) => {
-  const { fullName, email, phoneNumber, status } = userInfo;
-
   return (
     <div id={styles['admin-settings']}>
-      <Container className={styles['container-sect']}>
-        <h3>
-          Account Information{' '}
-          <Link className="btn btn-warning" to="/settings/edit" state={userInfo}>
-            Edit
-          </Link>
-        </h3>
-        <Row>
-          <Col xs sm md={{ span: 5, offset: 1 }}>
-            <h5>Name</h5>
-            <p className={styles['section-val']}>{fullName}</p>
-          </Col>
-          <Col md={5} className="d-none d-md-block">
-            <h5>Status</h5>
-            <p
-              style={{ color: status === 'Active' ? '#28A745' : '#6c757d' }}
-              className={`${styles['section-val']} ${styles['mt-status']}`}
-            >
-              {status}
-            </p>
-          </Col>
-        </Row>
-        <Row>
-          <Col md={{ offset: 1 }}>
-            <h5>Email</h5>
-            <p className={styles['section-val']}>{email}</p>
-          </Col>
-        </Row>
-        <Row>
-          <Col md={{ offset: 1 }}>
-            <h5>Phone Number</h5>
-            <p className={styles['section-val']}>{phoneNumber}</p>
-          </Col>
-        </Row>
-        <Row className="d-block d-sm-block d-md-none">
-          <Col>
-            <h5>Status</h5>
-            <p
-              style={{ color: status === 'Active' ? '#28A745' : '#6c757d' }}
-              className={`${styles['section-val']} ${styles['mt-status']}`}
-            >
-              {status}
-            </p>
-          </Col>
-        </Row>
-      </Container>
-      <Container className={styles['container-sect']}>
-        <h3>Password</h3>
-        <Row>
-          <Col md={{ offset: 1 }}>
-            <h5>
-              Current Password **********
-              <Link
-                id={styles['change-pwd-btn']}
-                className="btn btn-warning"
-                to="/settings/edit"
-                state={userInfo}
-              >
-                Change Password
-              </Link>
-            </h5>
-          </Col>
-        </Row>
-      </Container>
-      {/* TODO: Accessibility Section */}
-      {/* <Container className={styles['container-sect']}>
-        <h3>Accessibility</h3>
-        <Row>
-          <Col md={{ offset: 1 }}>
-            <h5>Text Size</h5>
-            <p className={styles['section-val']}>Default Text Larger Text</p>
-          </Col>
-        </Row>
-      </Container> */}
+      <AccountInformationView userInfo={userInfo} />
+      <PasswordView userInfo={userInfo} />
     </div>
   );
-};
-
-AdminView.defaultProps = {
-  userInfo: {
-    fullName: '',
-    email: '',
-    phoneNumber: '',
-    status: '',
-  },
 };
 
 AdminView.propTypes = {
@@ -102,7 +19,7 @@ AdminView.propTypes = {
     email: PropTypes.string,
     phoneNumber: PropTypes.string,
     status: PropTypes.string,
-  }),
+  }).isRequired,
 };
 
 export default AdminView;
