+++ src/pages/settings/teacherView.jsx
@@ -1,60 +1,12 @@
 import React from 'react';
 import { PropTypes } from 'prop-types';
-import { Container, Row, Col } from 'react-bootstrap';
-import { Link } from 'react-router-dom';
 import styles from './settings.module.css';
+import AccountInformationView from './accountInformationView';
 
 const TeacherView = ({ userInfo }) => {
-  const { fullName, email, phoneNumber, status } = userInfo;
-
   return (
     <div id={styles['mt-settings']}>
-      <Container className={styles['container-sect']}>
-        <h3>
-          Account Information{' '}
-          <Link className="btn btn-warning" to="/settings/edit" state={userInfo}>
-            Edit
-          </Link>
-        </h3>
-        <Row>
-          <Col xs sm md={{ span: 5, offset: 1 }}>
-            <h5>Name</h5>
-            <p className={styles['section-val']}>{fullName}</p>
-          </Col>
-          <Col md={5} className="d-none d-md-block">
-            <h5>Status</h5>
-            <p
-              style={{ color: status === 'Active' ? '#28A745' : '#6c757d' }}
-              className={`${styles['section-val']} ${styles['mt-status']}`}
-            >
-              {status}
-            </p>
-          </Col>
-        </Row>
-        <Row>
-          <Col md={{ offset: 1 }}>
-            <h5>Email</h5>
-            <p className={styles['section-val']}>{email}</p>
-          </Col>
-        </Row>
-        <Row>
-          <Col md={{ offset: 1 }}>
-            <h5>Phone Number</h5>
-            <p className={styles['section-val']}>{phoneNumber}</p>
-          </Col>
-        </Row>
-        <Row className="d-block d-sm-block d-md-none">
-          <Col>
-            <h5>Status</h5>
-            <p
-              style={{ color: status === 'Active' ? '#28A745' : '#6c757d' }}
-              className={`${styles['section-val']} ${styles['mt-status']}`}
-            >
-              {status}
-            </p>
-          </Col>
-        </Row>
-      </Container>
+      <AccountInformationView userInfo={userInfo} />
       {/* TODO: Accessibility section */}
       {/* <Container className={styles['container-sect']}>
         <h3>Accessibility</h3>
